Working Debugger list(.deb) Saving and Loading finished.
It's even less of a hassle to use now than it is to use in FCEUXDSP. Debugger/watchpoints automatically reload on game change, no phantom information traversing .deb files, and it's compatible with the FCEUXDSP .deb files.

Took a long time to write & debug the code, so there may still be errors I didn't catch.

git-svn-id: svn://svn.code.sf.net/p/fceultra/code/fceu@1437 cd624aef-3799-4ffa-999e-25a118dd8ffa