added --periodicsaves command line argument
in sdl.cpp:DoFun(), it will save to the state specified by
--savestate periodically.  The interval is currently decided by
a #define in sdl.h. I have it this way because you have to mod
by the interval, and having this be a compile time value lets 
the compiler optimize out the modulo operation.  Since this check
will happen so frequently I thought this was better.  However, if
this is better suited to something configurable, I am willing to
change it.


git-svn-id: svn://svn.code.sf.net/p/fceultra/code/fceu@2934 cd624aef-3799-4ffa-999e-25a118dd8ffa